requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
pillow==10.4.0